+++ CHANGELOG.md
@@ -9,6 +9,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2022-10-11
+
+### Added
+
+- Added `lfiles` module.
+- Added functions `llists:enumerate/1`, `llists:enumerate/2`, `llists:uniq/1`
+  and `llists:uniq/2`.
+
+### Deprecated
+
+- The `llist_utils:enumerate/1` function is now replaced by
+  `llists:enumerate/1` and should no longer be used.
+
 ## [1.1.0] - 2019-4-19
 
 ### Added
@@ -25,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[unreleased]: https://github.com/jkrukoff/llists/compare/v1.1.0...HEAD
+[unreleased]: https://github.com/jkrukoff/llists/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/jkrukoff/llists/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/jkrukoff/llists/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/jkrukoff/llists/releases/tag/v1.0.0